Global Shader Example
An example scene utilizing a “global” shader attached directly to the
MainCamera.

One shader is set up to show a vertex shader that changes the position
of vertices prior to rendering by panning them up/down along a moving
sine wave. Shapes made up of few vertices (e.g. the lower platforms) do
not “wave” properly because only their corners are interpolated. An
additional platform near the top is made up of finer grained geometry
and therefore “waves” more correctly.

The other example shader calculates the distance from points to the
origin (center of the screen) and determines the color drawn based on
this distance, either with a hard cutoff, or with a smooth gradient.

Shaders are attached to the MainCamera through the “CameraShaderScript”
script, which has a publicly settable option for a Shader object.