Camera script + new Prefabs
In the testing grounds script folder is the cameraFollow script. Drag it
onto your camera object or select it from a list, and then set the MAX
and MIN for x and y, mine are 999 or -999 currently. You can then adjust
the smoothing/ect to suit your needs.

New Prefabs: Pressure Plates, Crates, and Barrels