Bow trajectory working
Note to self: rename "Tragectory Script" at some point, since whoever
named it misspelled it. Also, we have a semi-working trajectory script,
though it only works with the bullet, not with the arrow. I think it may
have something to do with Angular Drag, but I haven't gotten it fixed
yet.