DO NOT CHANGE THE .gitignore FILE
The .gitignore file keeps getting “*.meta” added to the end of it —
this is BAD, and is part of the root cause of our “script explosions” —
the meta files are important and need to be sync’d across machines.
Please do not edit or make any changes to the .gitignore file!