Just Kidding
This commit undoes almost everything in my previous commit, because I
finally looked at the artemismove script and realized that it already
takes care of the death fade. It just required a change to the health
script to take full effect.

I also added a scene change checkpoint that activates when the hind
dies.